Move FollowRequestListAdapter into ui folder where MainActivity is, since MainActivity is the one that handles Follow Requests, not profile.
