+++ code/app/src/main/java/com/cmput301f21t26/habittracker/ui/FollowRequestListAdapter.java
@@ -1,4 +1,4 @@
-package com.cmput301f21t26.habittracker.ui.profile;
+package com.cmput301f21t26.habittracker.ui;
 
 import android.content.Context;
 import android.view.LayoutInflater;
